Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -13,7 +13,7 @@ int main(){
         printf("%d",x*y);
     }
     else if(z=='/'){
-        printf("%d",x/y);
+        printf("%d",(int)x/y);
     }
     else{
         printf("error");
